# IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "2.4.0",
    "title" : "CustomerManagement"
  },
  "host" : "serverRoot",
  "basePath" : "/customerManagement",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/customer/{id}/paymentMethod" : {
      "patch" : {
        "tags" : [ "paymentMethod" ],
        "summary" : "Updates paymentMethod of a 'Customer' by Id",
        "operationId" : "patchCustomerPaymentMethod",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "description" : "token to be passed as a header linked to the user whose identity is requested",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Customer",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "PaymentMethodRef",
          "description" : "PaymentMethodRef",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PaymentMethodRef"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/PaymentMethodRef"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "Conflict",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "PaymentMethodRef" : {
      "type" : "object",
      "required" : [ "href", "id" ],
      "properties" : {
        "href" : {
          "type" : "string",
          "description" : "Reference of the payment mean"
        },
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the payment mean"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the payment mean"
        }
      },
      "description" : "PaymentMethod reference. A payment method defines a specific mean of payment (e.g direct debit).",
      "example" : {
        "name" : "name",
        "href" : "href",
        "id" : "id"
      }
    },
...
    "Error" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "description" : "An integer coding the error type. This is given to caller so he can translate them if required."
        },
        "message" : {
          "type" : "string",
          "description" : "A short localized string that describes the error."
        },
        "description" : {
          "type" : "string",
          "description" : "(optional) A long localized error description if needed. It can contain precise information about which parameter is missing, or what are the identifier acceptable values."
        },
        "infoURL" : {
          "type" : "string",
          "description" : "(optional) A URL to online documentation that provides more information about the error."
        }
      }
    }
  }
}